New: legend exported, control hasDangerBorder
